Modify: User into UserView
